% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xll_round.R
\name{xl_round}
\alias{xl_round}
\title{Round Numbers Like Excel}
\usage{
xl_round(x, digits = 0)
}
\arguments{
\item{x}{numeric vector.}

\item{digits}{integer indicating the number of decimal places.}
}
\description{
Mimics the default rounding in Excel.
}
\examples{
xl_round(.5 + -2:4)
}
